Bump up warnings in dev mode.  Fix myriad of warnings from compiling in dev mode
Fix bug with counter not incremented.

Fix potential buffer overrun with vint decoding

git-svn-id: https://svn.mcs.anl.gov/repos/exm/sfw/lb/trunk@10112 dc4e9af1-7f46-4ead-bba6-71afc04862de

+++ code/src/notifications.h
@@ -74,7 +74,7 @@ typedef struct {
   // (may be NULL)
   void **to_free;
   int to_free_length;
-  int to_free_size; // Allocated length
+  size_t to_free_size; // Allocated length
 } adlb_notif_t;
 
 #define ADLB_NO_NOTIF_RANKS { .count = 0, .notifs = NULL }
